# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Smart Platform Authentication API
  version: 0.3.0
  description: Email identity contract for SmartTro UC-01, UC-02, and UC-03.
paths:
  /auth/signup/otp/request:
    post:
      summary: Request a registration OTP by email
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                previousAttemptId:
                  type: string
                  format: uuid
                  description: Live attempt being superseded only when changing normalized email.
      responses:
        '202':
          description: Generic response for both registered and unregistered emails
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpAccepted'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          description: Email provider is unavailable; production fails closed
  /auth/signup/otp/verify:
    post:
      summary: Verify the six-digit registration OTP
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, attemptId, code]
              properties:
                email:
                  type: string
                  format: email
                attemptId:
                  type: string
                  format: uuid
                code:
                  type: string
                  pattern: '^\d{6}$'
      responses:
        '200':
          description: OTP verified; no OTP, password, or token is returned
          content:
            application/json:
              schema:
                type: object
                required: [verified, attemptId, expiresInSeconds]
                properties:
                  verified:
                    const: true
                  attemptId:
                    type: string
                    format: uuid
                  expiresInSeconds:
                    type: integer
                    minimum: 1
        '401':
          description: Generic invalid, expired, exhausted, or already-used OTP
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/signup/complete:
    post:
      summary: Create the verified renter account and return to Sign In
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, attemptId, code, password, displayName]
              properties:
                email:
                  type: string
                  format: email
                attemptId:
                  type: string
                  format: uuid
                code:
                  type: string
                  pattern: '^\d{6}$'
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
                  description: Must include lowercase, uppercase, number, and special character.
                displayName:
                  type: string
                  minLength: 1
                  maxLength: 160
                  description: Required full name, normalized by trimming and collapsing repeated whitespace.
                phone:
                  type: string
                  nullable: true
                  pattern: '^\+[1-9]\d{7,14}$'
                  description: Optional contact data; not a login identifier.
      responses:
        '201':
          description: Account created; no access or refresh token is issued
          content:
            application/json:
              schema:
                type: object
                required: [created, next]
                properties:
                  created:
                    const: true
                  next:
                    const: sign_in
        '400':
          description: Generic invalid, expired, consumed, duplicate, or raced signup attempt
  /auth/login:
    post:
      summary: Sign in with normalized email and password
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  maxLength: 128
      responses:
        '200':
          description: Access token (15 minutes) and rotating refresh token (30 days)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionTokens'
        '401':
          description: Generic INVALID_CREDENTIALS response
        '403':
          description: ACCOUNT_UNVERIFIED or ACCOUNT_INACTIVE after valid credentials
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/password/recovery/request:
    post:
      summary: Request a password recovery OTP without revealing account existence
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailInput'
      responses:
        '202':
          description: Neutral response for eligible, nonexistent, unverified, and social-only accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordRecoveryAccepted'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          description: Email delivery is not configured for the environment
  /auth/password/recovery/verify:
    post:
      summary: Verify a recovery OTP and issue a context-bound reset token
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, challengeId, code]
              properties:
                email:
                  type: string
                  format: email
                challengeId:
                  type: string
                  format: uuid
                code:
                  type: string
                  pattern: '^\d{6}$'
      responses:
        '200':
          description: OTP consumed and a single-use reset token issued for 10 minutes
          content:
            application/json:
              schema:
                type: object
                required: [verified, resetToken, expiresInSeconds]
                properties:
                  verified:
                    const: true
                  resetToken:
                    type: string
                    writeOnly: true
                  expiresInSeconds:
                    type: integer
                    example: 600
        '401':
          description: PASSWORD_RECOVERY_OTP_INVALID
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/password/recovery/reset:
    post:
      summary: Reset the password and revoke every existing session
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resetToken, newPassword, confirmPassword]
              properties:
                resetToken:
                  type: string
                  writeOnly: true
                  minLength: 32
                newPassword:
                  type: string
                  writeOnly: true
                  minLength: 8
                  maxLength: 128
                  description: Must include lowercase, uppercase, number, and special character.
                confirmPassword:
                  type: string
                  writeOnly: true
                  minLength: 8
                  maxLength: 128
      responses:
        '200':
          description: Password reset; no session is created
          content:
            application/json:
              schema:
                type: object
                required: [reset, next, email]
                properties:
                  reset:
                    const: true
                  next:
                    const: sign_in
                  email:
                    type: string
                    format: email
        '400':
          description: Password policy, confirmation, or unchanged-password error
        '401':
          description: PASSWORD_RESET_TOKEN_INVALID
  /auth/token/refresh:
    post:
      summary: Rotate a refresh token
      responses:
        '200':
          description: Replacement token pair
  /auth/logout:
    post:
      summary: Revoke the session represented by a refresh token
      responses:
        '204':
          description: Session revoked; repeated calls are safe
  /auth/logout/all:
    post:
      summary: Revoke every active session for the authenticated account
      security:
        - bearerAuth: []
      responses:
        '204':
          description: All sessions revoked
  /profile:
    get:
      summary: Read the authenticated renter's UC-04 profile and active rentals
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Subject-scoped profile; response is private and must not be cached.
          headers:
            Cache-Control:
              schema:
                const: private, no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenterProfileRead'
        '401':
          description: INVALID_SESSION for expired, revoked, inactive, or invalid sessions
        '403':
          description: ROLE_FORBIDDEN when the actor is not a renter
        '409':
          description: PROFILE_INCOMPLETE when the authenticated renter has no usable email
        '503':
          description: PROFILE_UNAVAILABLE for profile/contract dependency failures or incomplete active-rental relations
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    DeviceId:
      name: X-Device-Id
      in: header
      required: false
      schema:
        type: string
        maxLength: 255
      description: Stable client-generated identifier used only as an abuse-control key.
  schemas:
    EmailInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
    OtpAccepted:
      type: object
      required: [accepted, attemptId, expiresInSeconds, resendAfterSeconds]
      properties:
        accepted:
          const: true
        attemptId:
          type: string
          format: uuid
        expiresInSeconds:
          type: integer
          example: 600
        resendAfterSeconds:
          type: integer
          example: 60
    PasswordRecoveryAccepted:
      type: object
      required: [accepted, message, challengeId, expiresInSeconds, resendAfterSeconds]
      properties:
        accepted:
          const: true
        message:
          const: Nếu email tồn tại, mã xác thực đã được gửi.
        challengeId:
          type: string
          format: uuid
          description: Opaque correlation value; it does not prove that the account exists.
        expiresInSeconds:
          type: integer
          example: 600
        resendAfterSeconds:
          type: integer
          example: 60
    SessionTokens:
      type: object
      required:
        [accessToken, refreshToken, tokenType, accessExpiresInSeconds, refreshExpiresInSeconds]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          const: Bearer
        accessExpiresInSeconds:
          type: integer
          example: 900
        refreshExpiresInSeconds:
          type: integer
          example: 2592000
    RenterProfileRead:
      type: object
      required: [profile, rentals]
      additionalProperties: false
      properties:
        profile:
          type: object
          required: [displayName, email, phone, avatar]
          additionalProperties: false
          properties:
            displayName:
              type: string
              description: users.display_name, or the approved legacy fallback Người dùng SmartTrọ.
            email:
              type: string
              format: email
              description: Required users.email; blank or missing values return PROFILE_INCOMPLETE.
            phone:
              type: [string, 'null']
              description: Raw users.phone for authenticated FE-side masking.
            avatar:
              type: 'null'
              description: Always null until an approved avatar source exists.
        rentals:
          type: array
          items:
            type: object
            required: [contractId, room, property, propertyAddress, expiresAt, signedAt]
            additionalProperties: false
            properties:
              contractId:
                type: string
                format: uuid
                description: contracts.id.
              room:
                type: string
                description: rooms.name, including soft-deleted or inactive rooms while the contract remains active.
              property:
                type: string
                description: properties.name, including soft-deleted or inactive properties while the contract remains active.
              propertyAddress:
                type: object
                additionalProperties: true
                description: properties.address. This source data is supplied for Home display and must not be derived from properties.name.
              expiresAt:
                type: [string, 'null']
                format: date
                description: contracts.ends_on; null represents an open-ended active contract.
              signedAt:
                type: string
                format: date-time
                description: contracts.renter_signed_at, with contracts.activated_at only as the approved legacy fallback.
  responses:
    RateLimited:
      description: AUTH_RATE_LIMITED with retryAfterSeconds